Initial commit for build_schema_from_type_definitions
