+++ ariadne/__init__.py
@@ -1,11 +1,12 @@
 from .add_resolve_functions_to_schema import add_resolve_functions_to_schema
+from .build_schema_from_type_definitions import build_schema_from_type_definitions
 from .default_resolver import default_resolver
-from .make_executable_schema import build_schema, make_executable_schema
+from .make_executable_schema import make_executable_schema
 
 
 __all__ = [
     "add_resolve_functions_to_schema",
-    "build_schema",
+    "build_schema_from_type_definitions",
     "default_resolver",
     "make_executable_schema",
 ]
